examples/tutorials/basic/mongoengine/requirements.txt
starlette-admin
uvicorn
mongoengine
pillow
